Refactor Image to have sub-types
