+++ src/Blitter.Bits/Cubemaps.cs
@@ -238,10 +238,21 @@ public static class Cubemaps
         int size = c.Size;
         int x = Math.Clamp((int)((u + 1f) * 0.5f * size), 0, size - 1);
         int y = Math.Clamp((int)((v + 1f) * 0.5f * size), 0, size - 1);
-        return c.GetFace(face).GetPixel(x, y);
+        return AsBitmap(c.GetFace(face)).GetPixel(x, y);
     }
 
-    private static Image BakeFace(int size, CubeFace face, Func<CubeFace, Vector3, Color> shade)
+    // CPU sampling needs raw pixels. Unwrap a mip chain to its base
+    // level; anything else (e.g. a future GPU-only image) can't be
+    // sampled on the CPU.
+    private static BitmapImage AsBitmap(Image image) => image switch
+    {
+        BitmapImage bitmap => bitmap,
+        MipmappedImage mipmapped => AsBitmap(mipmapped.Base),
+        _ => throw new NotSupportedException(
+            $"Cubemap face image of type {image.GetType().Name} cannot be sampled on the CPU."),
+    };
+
+    private static BitmapImage BakeFace(int size, CubeFace face, Func<CubeFace, Vector3, Color> shade)
     {
         var image = Image.Create(size, size, PixelFormat.ABGR8888);
         float inv = 2f / size;
