+++ src/Blitter/SkiaSharpExtensions.cs
@@ -5,7 +5,7 @@ namespace Blitter;
 
 public static class SkiaSharpExtensions
 {
-    extension(Image image)
+    extension(BitmapImage image)
     {
         /// <summary>
         /// Copies pixels from <paramref name="bitmap"/> into <paramref name="image"/>.
@@ -122,7 +122,7 @@ public static class SkiaSharpExtensions
         /// </summary>
         /// <param name="sigmaX">Horizontal blur standard deviation, in pixels.</param>
         /// <param name="sigmaY">Vertical blur standard deviation, in pixels.</param>
-        public Image Blur(float sigmaX, float sigmaY)
+        public BitmapImage Blur(float sigmaX, float sigmaY)
         {
             using var filter = SKImageFilter.CreateBlur(sigmaX, sigmaY);
             return ApplyImageFilter(image, filter, padLeft: 0, padTop: 0, padRight: 0, padBottom: 0);
@@ -133,7 +133,7 @@ public static class SkiaSharpExtensions
         /// uniform Gaussian blur applied.
         /// </summary>
         /// <param name="sigma">Blur standard deviation, in pixels.</param>
-        public Image Blur(float sigma) => image.Blur(sigma, sigma);
+        public BitmapImage Blur(float sigma) => image.Blur(sigma, sigma);
 
         /// <summary>
         /// Returns a new image with a drop shadow composited under the
@@ -147,7 +147,7 @@ public static class SkiaSharpExtensions
         /// <param name="sigmaX">Horizontal shadow blur, in pixels.</param>
         /// <param name="sigmaY">Vertical shadow blur, in pixels.</param>
         /// <param name="color">Shadow color (alpha included).</param>
-        public Image DropShadow(float dx, float dy, float sigmaX, float sigmaY, Blitter.Color color)
+        public BitmapImage DropShadow(float dx, float dy, float sigmaX, float sigmaY, Blitter.Color color)
         {
             // Pad enough to fit the shadow's blurred footprint plus
             // the offset on whichever side the shadow falls.
@@ -167,7 +167,7 @@ public static class SkiaSharpExtensions
         /// Returns a new image the same size as the source converted
         /// to grayscale via the standard luminance weights.
         /// </summary>
-        public Image Grayscale()
+        public BitmapImage Grayscale()
         {
             // Rec. 601 luma coefficients applied to R, G, B; alpha untouched.
             ReadOnlySpan<float> matrix =
@@ -188,7 +188,7 @@ public static class SkiaSharpExtensions
         /// (treated as 0..1). Alpha is multiplied like the others, so
         /// a fully opaque tint preserves the source alpha.
         /// </summary>
-        public Image Tint(Blitter.Color tint)
+        public BitmapImage Tint(Blitter.Color tint)
         {
             float r = tint.R / 255f;
             float g = tint.G / 255f;
@@ -213,7 +213,7 @@ public static class SkiaSharpExtensions
         /// </summary>
         /// <param name="radiusX">Horizontal dilate radius, in pixels.</param>
         /// <param name="radiusY">Vertical dilate radius, in pixels.</param>
-        public Image Dilate(int radiusX, int radiusY)
+        public BitmapImage Dilate(int radiusX, int radiusY)
         {
             using var filter = SKImageFilter.CreateDilate(radiusX, radiusY);
             return ApplyImageFilter(image, filter, radiusX, radiusY, radiusX, radiusY);
@@ -226,7 +226,7 @@ public static class SkiaSharpExtensions
         /// </summary>
         /// <param name="radiusX">Horizontal erode radius, in pixels.</param>
         /// <param name="radiusY">Vertical erode radius, in pixels.</param>
-        public Image Erode(int radiusX, int radiusY)
+        public BitmapImage Erode(int radiusX, int radiusY)
         {
             using var filter = SKImageFilter.CreateErode(radiusX, radiusY);
             return ApplyImageFilter(image, filter, 0, 0, 0, 0);
@@ -239,7 +239,7 @@ public static class SkiaSharpExtensions
         /// appears at (<c>max(0,-dx)+dx</c>, <c>max(0,-dy)+dy</c>) in
         /// the returned image.
         /// </summary>
-        public Image Offset(float dx, float dy)
+        public BitmapImage Offset(float dx, float dy)
         {
             int padLeft   = (int)MathF.Ceiling(MathF.Max(0f, -dx));
             int padTop    = (int)MathF.Ceiling(MathF.Max(0f, -dy));
@@ -261,7 +261,7 @@ public static class SkiaSharpExtensions
         /// <param name="kernel">Row-major kernel weights.</param>
         /// <param name="gain">Output multiplier applied after convolution. Defaults to 1.</param>
         /// <param name="bias">Constant added to each channel after the gain. Defaults to 0.</param>
-        public Image Convolve(int kernelWidth, int kernelHeight, ReadOnlySpan<float> kernel, float gain = 1f, float bias = 0f)
+        public BitmapImage Convolve(int kernelWidth, int kernelHeight, ReadOnlySpan<float> kernel, float gain = 1f, float bias = 0f)
         {
             if (kernel.Length != kernelWidth * kernelHeight)
                 throw new ArgumentException(
@@ -298,7 +298,7 @@ public static class SkiaSharpExtensions
         /// photographic content; <see cref="ImageSampling.Nearest"/>
         /// preserves crisp pixels for pixel-art / debug zoom uses.
         /// </param>
-        public Image Magnify(Rect lens, float zoom, float inset = 0f, ImageSampling sampling = ImageSampling.Linear)
+        public BitmapImage Magnify(Rect lens, float zoom, float inset = 0f, ImageSampling sampling = ImageSampling.Linear)
         {
             var lensRect = new SKRect(lens.X, lens.Y, lens.X + lens.Width, lens.Y + lens.Height);
             var skSampling = sampling == ImageSampling.Nearest
@@ -341,7 +341,7 @@ public static class SkiaSharpExtensions
             _ => FastPath.None,
         };
 
-    private static unsafe bool TryFastCopyFromBitmap(SKBitmap bitmap, Image image, int width, int height)
+    private static unsafe bool TryFastCopyFromBitmap(SKBitmap bitmap, BitmapImage image, int width, int height)
     {
         if (image.BytesPerPixel != 4)
             return false;
@@ -384,7 +384,7 @@ public static class SkiaSharpExtensions
         return true;
     }
 
-    private static unsafe bool TryFastCopyToBitmap(Image image, SKBitmap bitmap, int width, int height)
+    private static unsafe bool TryFastCopyToBitmap(BitmapImage image, SKBitmap bitmap, int width, int height)
     {
         if (image.BytesPerPixel != 4)
             return false;
@@ -438,7 +438,7 @@ public static class SkiaSharpExtensions
         }
     }
 
-    private static void DrawCanvasCore(Image image, Blitter.Color? backgroundColor, Action<SKCanvas> drawAction)
+    private static void DrawCanvasCore(BitmapImage image, Blitter.Color? backgroundColor, Action<SKCanvas> drawAction)
     {
         ArgumentNullException.ThrowIfNull(drawAction);
         ObjectDisposedException.ThrowIf(image.IsDisposed, image);
@@ -484,8 +484,8 @@ public static class SkiaSharpExtensions
         image.CopyFromBitmap(bitmap);
     }
 
-    private static Image ApplyImageFilter(
-        Image source, SKImageFilter filter,
+    private static BitmapImage ApplyImageFilter(
+        BitmapImage source, SKImageFilter filter,
         int padLeft, int padTop, int padRight, int padBottom)
     {
         ObjectDisposedException.ThrowIf(source.IsDisposed, source);
@@ -538,7 +538,7 @@ public static class SkiaSharpExtensions
         /// Creates a new <see cref="Image"/> the same size as the bitmap
         /// and copies its pixels into it.
         /// </summary>
-        public Image ToImage(Blitter.PixelFormat? format = null)
+        public BitmapImage ToImage(Blitter.PixelFormat? format = null)
         {
             ArgumentNullException.ThrowIfNull(bitmap);
 
@@ -587,7 +587,7 @@ public static class SkiaSharpExtensions
     /// </summary>
     private sealed class CanvasDrawState : IDisposable
     {
-        public Image? Image;
+        public BitmapImage? Image;
         public SKBitmap? Bitmap;
         public SKCanvas? Canvas;
